Move tests into standalone file
